functional/functional.go
package functional

import (
	"iter"
)

// Stream represents a sequence of integers that can be processed using functional operations.
type Stream struct {
	// Add your fields here
}

// NewStream creates a new Stream from a slice of integers.
func NewStream(values []int) *Stream {
	// TODO: Implement this function
	return nil
}

// Map applies the function f to each element in the stream and returns a new stream with the results.
func (s *Stream) Map(f func(int) int) *Stream {
	// TODO: Implement this function
	return nil
}

// Filter returns a new stream containing only the elements that satisfy the predicate.
func (s *Stream) Filter(predicate func(int) bool) *Stream {
	// TODO: Implement this function
	return nil
}

// Take returns a new stream containing at most the first n elements.
func (s *Stream) Take(n int) *Stream {
	// TODO: Implement this function
	return nil
}

// Drop returns a new stream with the first n elements removed.
func (s *Stream) Drop(n int) *Stream {
	// TODO: Implement this function
	return nil
}

// Iterate returns an iterator over the elements in the stream.
func (s *Stream) Iterate() iter.Seq[int] {
	// TODO: Implement this function
	return nil
}

// FoldLeft reduces the stream to a single value using the function f, processing from left to right.
func (s *Stream) FoldLeft(f func(int, int) int) int {
	// TODO: Implement this function
	return 0
}

// ForEach applies the function f to each element in the stream.
func (s *Stream) ForEach(f func(int)) {
	// TODO: Implement this function
}
